third_party/WebKit/LayoutTests/editing/inserting/insert-html-image-with-align.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() =>
  assert_selection(
    '<div contenteditable>foobar|baz</div>',
    'InsertHTML <img src=../../resources/abe.png align=right>',
    '<div contenteditable>foobar<img align="right" src="../../resources/abe.png">|baz</div>'),
  'InsertHTML image with align');
</script>
